test(helper): add helpers directory
