+++ src/server/__tests__/helpers/mocks.ts
@@ -1,17 +1,15 @@
 import { HttpError } from "http-errors";
 import { createRequest, createResponse } from "node-mocks-http";
-import { Permission } from "../../../../shared/api/entities";
-import { sessions, users } from "../../../session/__tests__/helpers";
+import { Permission } from "../../../shared/api/entities";
+import { users } from "./entities";
 
 export const createHttpMocks = (permission: Permission) => {
   const req = createRequest();
   const res = createResponse();
 
-  req.session = sessions[permission];
-
-  req.user = users[permission];
-  req.user.createdAt = new Date();
-  req.user.updatedAt = new Date();
+  req.session = {
+    user: users[permission]
+  } as any;
 
   const next = (error: HttpError) => {
     res.status(error.statusCode);
