+++ src/server/database/entities/__tests__/UserEntity.test.ts
@@ -1,5 +1,5 @@
 import { getManager } from "typeorm";
-import { TestDatabase } from "../../__tests__/helpers";
+import { TestDatabase } from "../../../__tests__/helpers";
 import { UserAccountEntity } from "../UserAccountEntity";
 import { UserConfigEntity } from "../UserConfigEntity";
 import { UserEntity } from "../UserEntity";
