+++ src/server/routes/__tests__/logout.test.ts
@@ -1,8 +1,6 @@
 import { serialize } from "cookie";
 import { sign } from "cookie-signature";
-import { testProcessEnv } from "../../__tests__/helpers";
-import { insertSessions, insertUsers, sessions } from "../../session/__tests__/helpers";
-import { TestServer } from "./helpers";
+import { insertSessions, insertUsers, sessions, testProcessEnv, TestServer } from "../../__tests__/helpers";
 
 const testServer = new TestServer();
