+++ src/server/routes/api/__tests__/users.test.ts
@@ -1,7 +1,6 @@
 import { SearchResponse } from "../../../../shared/api/response/search";
-import { TestDatabase } from "../../../database/__tests__/helpers";
+import { createHttpMocks, TestDatabase } from "../../../__tests__/helpers";
 import { GET } from "../users";
-import { createHttpMocks } from "./helpers";
 
 const testDatabase = new TestDatabase();
