+++ src/server/routes/api/exercises/__tests__/{id}.test.ts
@@ -1,12 +1,15 @@
 import { getManager } from "typeorm";
 import * as uuid from "uuid";
 import { EntityStore } from "../../../../../shared/api/response/get";
+import {
+  createHttpMocks,
+  insertExercise,
+  insertSessions,
+  insertUsers,
+  TestDatabase
+} from "../../../../__tests__/helpers";
 import { PathParams } from "../../../../api/operation";
-import { TestDatabase } from "../../../../database/__tests__/helpers";
 import { ExerciseEntity } from "../../../../database/entities";
-import { insertExercise } from "../../../../database/entities/__tests__/helpers";
-import { insertSessions, insertUsers } from "../../../../session/__tests__/helpers";
-import { createHttpMocks } from "../../__tests__/helpers";
 import { DELETE, GET } from "../{id}";
 
 const testDatabase = new TestDatabase();
