+++ src/server/routes/api/user-sessions/__tests__/{id}.test.ts
@@ -1,10 +1,15 @@
 import { getManager } from "typeorm";
 import * as uuid from "uuid";
 import { EntityStore } from "../../../../../shared/api/response/get";
+import {
+  createHttpMocks,
+  insertSessions,
+  insertUsers,
+  sessions,
+  TestDatabase,
+  users
+} from "../../../../__tests__/helpers";
 import { PathParams } from "../../../../api/operation";
-import { TestDatabase } from "../../../../database/__tests__/helpers";
-import { insertSessions, insertUsers, sessions, users } from "../../../../session/__tests__/helpers";
-import { createHttpMocks } from "../../__tests__/helpers";
 import { DELETE, GET } from "../{id}";
 
 const testDatabase = new TestDatabase();
